Moved NewestQCTracker into separate module
