+++ consensus/hotstuff/timeoutcollector/timeout_processor.go
@@ -3,12 +3,11 @@ package timeoutcollector
 import (
 	"errors"
 	"fmt"
-	"unsafe"
-
 	"go.uber.org/atomic"
 
 	"github.com/onflow/flow-go/consensus/hotstuff"
 	"github.com/onflow/flow-go/consensus/hotstuff/model"
+	"github.com/onflow/flow-go/consensus/hotstuff/tracker"
 	"github.com/onflow/flow-go/model/flow"
 	"github.com/onflow/flow-go/module/signature"
 )
@@ -33,45 +32,6 @@ func (t *accumulatedWeightTracker) Track(weight uint64) bool {
 	return t.done.CAS(false, true)
 }
 
-// NewestQCTracker is a helper structure which keeps track of the highest QC(by view)
-// in concurrency safe way.
-type NewestQCTracker struct {
-	newestQC *atomic.UnsafePointer
-}
-
-func NewNewestQCTracker() *NewestQCTracker {
-	tracker := &NewestQCTracker{
-		newestQC: atomic.NewUnsafePointer(unsafe.Pointer(nil)),
-	}
-	return tracker
-}
-
-// Track updates local state of NewestQC if the provided instance is newer(by view)
-// Concurrently safe
-func (t *NewestQCTracker) Track(qc *flow.QuorumCertificate) {
-	// to record the newest value that we have ever seen we need to use loop
-	// with CAS atomic operation to make sure that we always write the latest value
-	// in case of shared access to updated value.
-	for {
-		// take a snapshot
-		NewestQC := t.NewestQC()
-		// verify that our update makes sense
-		if NewestQC != nil && NewestQC.View >= qc.View {
-			return
-		}
-		// attempt to install new value, repeat in case of shared update.
-		if t.newestQC.CAS(unsafe.Pointer(NewestQC), unsafe.Pointer(qc)) {
-			return
-		}
-	}
-}
-
-// NewestQC returns the newest QC(by view) tracked.
-// Concurrently safe
-func (t *NewestQCTracker) NewestQC() *flow.QuorumCertificate {
-	return (*flow.QuorumCertificate)(t.newestQC.Load())
-}
-
 // TimeoutProcessor implements the hotstuff.TimeoutProcessor interface.
 // It processes timeout objects broadcast by other replicas of the consensus committee.
 // TimeoutProcessor collects TOs for one view, eventually when enough timeout objects are contributed
@@ -85,7 +45,7 @@ type TimeoutProcessor struct {
 	notifier         hotstuff.TimeoutCollectorConsumer
 	partialTCTracker accumulatedWeightTracker
 	tcTracker        accumulatedWeightTracker
-	newestQCTracker  *NewestQCTracker
+	newestQCTracker  *tracker.NewestQCTracker
 }
 
 var _ hotstuff.TimeoutProcessor = (*TimeoutProcessor)(nil)
@@ -122,7 +82,7 @@ func NewTimeoutProcessor(committee hotstuff.Replicas,
 			done:              *atomic.NewBool(false),
 		},
 		sigAggregator:   sigAggregator,
-		newestQCTracker: NewNewestQCTracker(),
+		newestQCTracker: tracker.NewNewestQCTracker(),
 	}, nil
 }
